reuse command to shutdown all cluster task instances
